Refactor DataSet class out of importers since they're all identical now
